Updated the function to youtube_trending_videos
Moved the input instructions within the youtube_trending_videos() that were set to be printed before each of the input functions. The instructions were initially placed above the input function and were supposed to instruct the user what to input. Since there are multiple tools that read code, a better place was found for the instructions and names inside the input function. This will help those who use tools similar to PyCharm, as the input instructions will be shown directly on the Jupyter Input Request window.